[POM-200] snack inprogress dirty snack wrks with success
