+++ src/app/app.component.html
@@ -1,6 +1,5 @@
 <app-site-header></app-site-header>
 
 <main [ngClass]="getContentClass()">
-  <button mat-raised-button color="primary" (click)="onklik()">kik</button>
   <router-outlet></router-outlet>
 </main>
