+++ src/app/give-help/material-aid-form/material-aid-form.component.html
@@ -1,5 +1,5 @@
 <mat-card>
-  <form (submitOffer)="handleSubmit()" #materialAidForm="ngForm">
+  <form (ngSubmit)="handleSubmit()" #materialAidForm="ngForm">
     <div class="row">
       <div class="col-md-12">
         <label>
@@ -36,7 +36,7 @@
             required
             #category="ngModel"
           >
-            <mat-option *ngFor="let it of categories" [value]="it.value">{{
+            <mat-option *ngFor="let it of CATEGORIES" [value]="it.value">{{
               it.value | translate | lowercase
             }}</mat-option>
           </mat-select>
@@ -46,7 +46,14 @@
         </mat-form-field>
       </div>
       <div class="col-md-6">
-        <label
+        <app-cities-search
+          [(ngModel)]="data.location"
+          name="location"
+          [placeholder]="'Rzeszów, podkarpackie'"
+          [label]="'LOCATION' | translate"
+          required
+        ></app-cities-search>
+        <!-- <label
           ><b>{{ "LOCATION" | translate }}:</b></label
         >
         <mat-form-field appearance="outline">
@@ -54,14 +61,14 @@
             matInput
             placeholder="{{ 'PLACEHOLDER_LOCATION' | translate }}"
             name="location"
-            [(ngModel)]="data.location"
+            [(ngModel)]="data.location.city"
             required
             #location="ngModel"
           />
           <mat-error>
             <app-field-error [model]="location"></app-field-error>
           </mat-error>
-        </mat-form-field>
+        </mat-form-field> -->
       </div>
     </div>
     <div class="row">
@@ -128,5 +135,6 @@
         </button>
       </div>
     </div>
+    <pre>{{ data | json }}</pre>
   </form>
 </mat-card>
